timeout/send dv/fail events are working.

+++ trunk/My_router/trunk/My_router/MyRIPMessage.h
@@ -11,17 +11,17 @@
 
 typedef struct
 {
-	int DestinationNETSubnet; //32 bit
-	int DestinationNETMask; //32 bit
-	int DestinationNETSubnetDistance; //32 bit
+	unsigned long DestinationNETSubnet; //32 bit
+	unsigned long DestinationNETMask; //32 bit
+	unsigned long DestinationNETSubnetDistance; //32 bit
 } DestinationProperties;
 
 typedef struct
 {
-	short length;
-	short protocolID;
-	int ConnectingNETMYIPSubnet; //32 bit
-	int ConnectingNETMYIPMask; //32 bit
+	unsigned short length;
+	unsigned short protocolID;
+	unsigned long ConnectingNETMYIPSubnet; //32 bit
+	unsigned long ConnectingNETMYIPMask; //32 bit
 	char SenderName[MAX_SENDER_NAME]; //64 bit
 	char ReceiverName[MAX_RECEIVER_NAME]; //64 bit
 	DestinationProperties dest[NUMBER_OF_DESTINATIONS];	//96 bit * NUMBER_OF_DESTINATIONS
